HOTFIX-EOI-RFI-COLUMN-UPDATE
